# apps/backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Helpmaton API",
    "version": "0.0.2",
    "description": "RESTful API for Helpmaton - AI agent workspace management and operations"
  },
  "servers": [
    {
      "url": "https://app.helpmaton.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:5173",
      "description": "Local Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token authentication using JWT access tokens. Use the format: Authorization: Bearer <access-token>. Access tokens are obtained via /api/user/generate-tokens or /api/user/refresh-token."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 401,
              "error": "Unauthorized",
              "message": "Authentication required"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 403,
              "error": "Forbidden",
              "message": "Insufficient permissions"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 400,
              "error": "Bad Request",
              "message": "Invalid request parameters"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 500,
              "error": "Internal Server Error",
              "message": "An unexpected error occurred"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "statusCode",
          "error"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "Workspace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Workspace ID"
          },
          "name": {
            "type": "string",
            "description": "Workspace name"
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "nullable": true
          },
          "permissionLevel": {
            "type": "integer",
            "description": "User's permission level in this workspace",
            "nullable": true
          },
          "creditBalance": {
            "type": "number",
            "description": "Current credit balance"
          },
          "currency": {
            "type": "string",
            "enum": [
              "usd",
              "eur",
              "gbp"
            ],
            "description": "Currency code"
          },
          "spendingLimits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpendingLimit"
            },
            "description": "Spending limits configuration"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          }
        }
      },
      "WorkspacesResponse": {
        "type": "object",
        "properties": {
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            }
          }
        }
      },
      "WorkspaceResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/Workspace"
          }
        ]
      },
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Workspace name"
          },
          "description": {
            "type": "string",
            "description": "Workspace description"
          },
          "currency": {
            "type": "string",
            "enum": [
              "usd",
              "eur",
              "gbp"
            ],
            "description": "Currency code",
            "default": "usd"
          }
        }
      },
...
      "SpendingLimit": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Limit amount"
          },
          "currency": {
            "type": "string",
            "enum": [
              "usd",
              "eur",
              "gbp"
            ],
            "description": "Currency code"
          },
          "period": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly"
            ],
            "description": "Time period"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/workspaces/{workspaceId}/spending-limits/{timeFrame}": {
      "put": {
        "summary": "Update workspace spending limit",
        "description": "Updates the amount for an existing spending limit for a specific time frame (daily, weekly, or monthly). The time frame must already exist. Requires WRITE permission or higher.",
        "tags": [
          "Spending Limits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "Workspace ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeFrame",
            "in": "path",
            "required": true,
            "description": "Time frame for the spending limit",
            "schema": {
              "type": "string",
              "enum": [
                "daily",
                "weekly",
                "monthly"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Spending limit amount"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace with updated spending limits",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Workspace"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete workspace spending limit",
        "description": "Removes a spending limit for a specific time frame (daily, weekly, or monthly). After deletion, there will be no limit for that time frame. Requires WRITE permission or higher.",
        "tags": [
          "Spending Limits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "Workspace ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeFrame",
            "in": "path",
            "required": true,
            "description": "Time frame for the spending limit",
            "schema": {
              "type": "string",
              "enum": [
                "daily",
                "weekly",
                "monthly"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace with updated spending limits",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Workspace"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/workspaces/{workspaceId}/members/{userId}": {
      "put": {
        "summary": "Update workspace member permission",
        "description": "Updates a member's permission level (1=READ, 2=WRITE, 3=OWNER). You cannot grant a permission level higher than your own. For example, a WRITE user cannot grant OWNER permissions. Requires WRITE permission or higher.",
        "tags": [
          "Workspace Members"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "Workspace ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "permissionLevel"
                ],
                "properties": {
                  "permissionLevel": {
                    "type": "integer",
                    "enum": [
                      1,
                      2,
                      3
                    ],
                    "description": "Permission level (1=READ, 2=WRITE, 3=OWNER)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member permission updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "userRef": {
                      "type": "string"
                    },
                    "permissionLevel": {
                      "type": "integer"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "410": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Remove workspace member",
        "description": "Removes a user from a workspace, revoking all their permissions. Cannot remove the last owner - a workspace must have at least one owner. This action cannot be undone. Requires OWNER permission.",
        "tags": [
          "Workspace Members"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "Workspace ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "User ID to remove",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "410": {
            "description": "Workspace or member not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/workspaces": {
      "post": {
        "summary": "Create a new workspace",
        "description": "Creates a new workspace for the authenticated user",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "get": {
        "summary": "List all workspaces",
        "description": "Returns all workspaces the authenticated user has access to",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}